load_accounting_db: move files to be public
PR #97 adds the ability to create a populated weighted
tree from a flux-accounting DB for testing purposes. That
functionality will be required moving forward to handle #63.

To make this easier, pull the load_accounting_db.* source files out
of the "test" directory and move them to a new directory "reader",
located under "fairness." Rename the load_accounting_db.* files to
data_reader_db.*.

Also move the unit test file and sample databases to the reader
directory.

Fixes #102
