+++ configure.ac
@@ -104,6 +104,8 @@ AC_CONFIG_FILES([Makefile
   src/fairness/Makefile
   src/fairness/weighted_tree/Makefile
   src/fairness/weighted_tree/test/Makefile
+  src/fairness/reader/Makefile
+  src/fairness/reader/test/Makefile
   src/fairness/print_hierarchy/Makefile
   src/bindings/Makefile
   src/bindings/python/Makefile
