+++ src/fairness/reader/data_reader_db.cpp
@@ -29,7 +29,7 @@
 #include <vector>
 #include <cerrno>
 
-#include "src/fairness/weighted_tree/test/load_accounting_db.hpp"
+#include "src/fairness/reader/data_reader_db.hpp"
 
 using namespace Flux::accounting;
